Move TextileImage to Components
