+++ App/Containers/AddCaptionScreen.tsx
@@ -8,7 +8,7 @@ import styles from '../SB/views/ThreadCreate/statics/styles'
 import ThreadSelect from '../SB/components/ThreadSelect'
 import UIActions from '../Redux/UIRedux'
 import PhotoViewingActions from '../Redux/PhotoViewingRedux'
-import TextileImage from '../../TextileImage'
+import TextileImage from '../Components/TextileImage'
 import {Photo, PhotoId, SharedImage, ThreadId, ThreadName} from '../Models/TextileTypes'
 import {RootAction, RootState} from '../Redux/Types'
 import {Dispatch} from 'redux'
